third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-not-block-import-in-import.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'">
<link id="target" rel="import" href="resources/importing-cors.html">
</head>
<body>
<pre id="result"></pre>
<!-- FIXME: The test should be renamed to csp-block-import-in-import.html -->
<script>
if (window.testRunner)
    testRunner.dumpAsText();

if (window.target.import.querySelector('#cors').import == null)
    result.innerHTML = "PASS\n";
else
    result.innerHTML = "FAIL: The import should be blocked.\n";
</script>
</body>
</html>
