third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-not-block-import-in-import-expected.txt

CONSOLE ERROR: Refused to load the script 'http://localhost:8000/htmlimports/resources/cors-basic.cgi' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

PASS
